Moved some general functions out of info_utils
