+++ celltk/utils/operation_utils.py
@@ -17,6 +17,49 @@ from celltk.utils._types import Mask
 
 # TODO: Add create lineage tree graph (maybe in plot_utils)
 
+def _nan_helper(y: np.ndarray) -> np.ndarray:
+    """Linear interpolation of nans in a 1D array."""
+    return np.isnan(y), lambda z: z.nonzero()[0]
+
+
+def nan_helper_2d(arr: np.ndarray) -> np.ndarray:
+    """Linear interpolation of nans along rows in 2D array.
+
+    TODO:
+        - Move to a more sensible util file
+    """
+    temp = np.zeros(arr.shape)
+    temp[:] = np.nan
+    for n, y in enumerate(arr.copy()):
+        nans, z = _nan_helper(y)
+        y[nans] = np.interp(z(nans), z(~nans), y[~nans])
+        temp[n, :] = y
+
+    return temp
+
+
+def nan_helper_1d(arr: np.ndarray) -> np.ndarray:
+    temp = arr.copy()
+    nans, z = _nan_helper(arr)
+    temp[nans] = np.interp(z(nans), z(~nans), arr[~nans])
+    return temp
+
+
+def get_split_idxs(arrays: Collection[np.ndarray], axis: int = 0) -> List[int]:
+    """
+    """
+    row_idxs = [s.shape[axis] for s in arrays]
+    split_idxs = [np.sum(row_idxs[:i + 1])
+                  for i in range(len(row_idxs))]
+    return split_idxs
+
+
+def split_array(array: np.ndarray, split_idxs: List[int], axis: int = 0) -> List[np.ndarray]:
+    """
+    """
+    return [n for n in np.split(array, split_idxs, axis=axis)
+            if n.shape[axis] > 0]
+
 
 def gray_fill_holes(labels: np.ndarray) -> np.ndarray:
     """
